+++ scripts/proxy.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+export HTTPS_PROXY=http://localhost:8080
+export HTTP_PROXY=http://localhost:8080
+export NODE_EXTRA_CA_CERTS=~/.mitmproxy/mitmproxy-ca-cert.pem
+